pyproject.toml
[project]
name = "zbq"
version = "0.1.5"
description = "A lightweight, wrapper around Google Cloud BigQuery with Polars integration. Simplifies querying and data ingestion with a unified interface, supporting read, write, insert, and delete operations on BigQuery tables."
readme = "README.md"
authors = [{ name = "FrocketGaming", email = "frocketgames@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery>=3.34.0",
    "google-cloud-bigquery-storage>=2.32.0",
    "google-cloud-storage>=3.1.1",
    "polars>=1.31.0",
    "pyarrow>=21.0.0",
    "tqdm>=4.67.1",
]

[project.scripts]
zbq = "zbq:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
